+++ ci/scripts/lint.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+set -e
+
+black /pantab
+isort /pantab/**/*.py
+mypy /pantab
+clang-format-11 /pantab/pantab/src/* -i --style=LLVM
+