+++ pantab/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.0.0"
+__version__ = "2.1.1"
 
 from tableauhyperapi import __version__ as hyperapi_version